Share tests across Ruby and CS repos
Reuse FakeGithubClient
Add a Gemfile.lock to test with
